Time: 3 ms (88.14%), Space: 12.6 MB (18.64%) - LeetHub

+++ 0576-out-of-boundary-paths/0576-out-of-boundary-paths.cpp
@@ -12,19 +12,19 @@ public:
         return ans;
     }
 
-    long dfs(int i, int j,int move,vector<vector<vector<long>>> & memo){
+    int dfs(int i, int j,int move,vector<vector<vector<long>>> & memo){
     if(move<0 )
         return 0;
     if(i < 0 || i>=M || j< 0 ||j >=N )
             return 1;
     
     if(move <= i && move <= j && move + i < M && move + j < N) return 0;
-    
+
     if(memo[i][j][move] != -1)
         return memo[i][j][move];
     
 
-    long total = 0;
+int total = 0;
     total  = (total+ dfs(i+1,j,move-1, memo))%MOD;
     total = (total+ dfs(i-1,j,move-1,memo))%MOD;
     total = (total+ dfs(i,j+1,move-1,memo))%MOD;
